requirements-dev.txt
-r requirements.txt  # incluye las de ejecución

black==25.1.0
isort==6.0.1
ruff==0.12.10
mypy==1.17.1
django-stubs==5.2.2
django-stubs-ext==5.2.2
types-PyYAML==6.0.12.20250822
click==8.2.1
colorama==0.4.6
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
mypy_extensions==1.1.0
typing_extensions==4.15.0
xhtml2pdf==0.2.10
